+++ ReleaseNotes.md
@@ -7,6 +7,7 @@ F# niggles part 4
 * Take account of F#'s habit of making a virtual call to the base type constructor in object types constructors.
 * Exempt F# code in modules, or where a `match` could equally be an `if` from `AvoidSwitchStatementsRule`, `match` being idiomatic and occasionally just happening to be on an explicit integral type
 * Exempt property backing fields for code like `member val LocalSource = false with get, set` from `AvoidUnneededFieldInitializationRule`
+* Exempt union cases (unsealed but not likely to be inherited) from `AvoidUnsealedUninheritedInternalTypeRule`